requirements.txt
solidpython
python-dotenv